requirements.txt
PyPDF2==3.0.1
mysql-connector-python==8.1.0
